requirements.txt
streamlit>=1.28.0
psycopg[binary,pool]>=3.1.0
databricks-sdk>=0.35.0
databricks-ai-bridge>=0.1.0
openai>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
requests>=2.31.0
mlflow>=2.10.0 